requirements.txt
yfinance
yfinance-cache
transformers
torch
requests
pandas
PyYAML
python-dotenv
mypy
flake8